Implement regular ping in lib-cache to check on distributed nodes.

+++ lib-cache/src/commonMain/kotlin/com/ustadmobile/libcache/distributed/model/DistributedCachePing.kt
@@ -3,7 +3,7 @@ package com.ustadmobile.libcache.distributed.model
 import java.nio.ByteBuffer
 
 data class DistributedCachePing(
-    override val id: Long,
+    override val id: Int,
     override val payload: ByteArray,
 ): DistributedCachePacket(), DistributedCacheWhatWithIdAndPayload {
